requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
datasets==1.18.4
dependency-injector==4.40.0
dill==0.3.5.1
dnspython==2.2.1
email-validator==1.2.1
emoji==1.7.0
fastapi==0.82.0
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.8.2
greenlet==1.1.3
h11==0.13.0
huggingface-hub==0.9.1
idna==3.3
joblib==1.1.0
Mako==1.2.2
mangum==0.15.1
MarkupSafe==2.1.1
multidict==6.0.2
multiprocess==0.70.13
mypy-extensions==0.4.3
numpy==1.23.2
packaging==21.3
pandas==1.4.4
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
pyarrow==9.0.0
pydantic==1.10.2
PyJWT==2.4.0
pyparsing==3.0.9
pysentimiento==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
responses==0.18.0
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
threadpoolctl==3.1.0
tokenizers==0.10.3
tomli==2.0.1
torch==1.12.1
tqdm==4.64.1
transformers==4.13.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.16.1
xxhash==3.0.0
yarl==1.8.1
